# src/openapi.yaml
openapi: 3.0.3
info:
  title: InvoiceSmart API
  version: 1.0.0
  description: |
    REST API powering the InvoiceSmart iOS app (and any future web client).
    Multi-tenant: every authenticated request is scoped to the JWT bearer's
    `user_id`. Cross-tenant reads return 404 (not 403) to avoid leaking
    existence.

    **Auth**: `POST /api/auth/login` returns a JWT, valid 7 days. Send as
    `Authorization: Bearer <jwt>` on every protected route. The `/health`
    and `/api/health` routes are unauthenticated and report DB
    connectivity for L7 probes.

    **Money** is `DECIMAL(12,2)` server-side and returned as a string
    (e.g. `"1500.00"`) — clients should `parseFloat` before arithmetic.
    Rates (`tax_rate`, `discount_rate`, `cis_rate`, `retention_rate`) are
    percentages, so `20` means 20% — never `0.20`.
servers:
  - url: https://api.invoicesmart.cortexbuildpro.com
    description: Production
  - url: http://127.0.0.1:3008
    description: Local dev / docker
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Email already registered }
  /api/auth/login:
    post:
      tags: [auth]
      summary: Exchange email+password for a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { description: Invalid credentials }
  /api/auth/me:
    get:
      tags: [auth]
      summary: Get current user's profile
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfile' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: User not found }
  /api/clients:
    get:
      tags: [clients]
      summary: List the current user's clients
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ClientList' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ClientInput' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Client' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/clients/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    get:
      tags: [clients]
      summary: Get one client
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Client' } } } }
        '400': { $ref: '#/components/responses/BadUuid' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [clients]
      summary: Update a client
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClientInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Client' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [clients]
      summary: Delete a client
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/invoices:
    get:
      tags: [invoices]
      summary: List invoices (optionally filtered by status)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [draft, sent, partial, paid] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceList' } } } }
    post:
      tags: [invoices]
      summary: Create an invoice (totals computed server-side)
      description: |
        Invoice numbers are unique **per user** (composite constraint
        `(user_id, invoice_number)`). Two tenants both reach `INV-0001`.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
  /api/invoices/next/number:
    get:
      tags: [invoices]
      summary: Compute the next invoice number for the current user
      parameters:
        - { in: query, name: prefix, schema: { type: string, default: INV } }
        - { in: query, name: autoIncrement, schema: { type: boolean, default: true } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { invoiceNumber: { type: string, example: INV-0042 } } }
  /api/invoices/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    get:
      tags: [invoices]
      summary: Get an invoice with line items and payments
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceFull' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [invoices]
      summary: Update an invoice (line items are replaced atomically)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [invoices]
      summary: Delete an invoice and its line items + transactions
      responses:
        '204': { description: Deleted }
  /api/invoices/{id}/send:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    post:
      tags: [invoices]
      summary: Mark invoice as sent (sets status=sent + sent_at)
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/invoices/{id}/paid:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    patch:
      tags: [invoices]
      summary: Force-mark an invoice as paid in full
      description: |
        Sets `amount_paid = total_amount`, `amount_due = 0`,
        `status = 'paid'`. Does NOT record a transaction. Use
        `/payments` instead when you want a transaction record.
      responses:
        '200': { description: OK }
  /api/invoices/{id}/pdf:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    get:
      tags: [invoices]
      summary: Get a signed URL for the invoice PDF
      description: Currently returns a stub URL — real PDF generation is on the roadmap.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { url: { type: string, format: uri } } }
  /api/invoices/{id}/payments:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    post:
      tags: [invoices]
      summary: Record a payment against an invoice
      description: |
        Atomically inserts a `transactions` row of type=`payment`,
        recalculates `amount_paid`, `amount_due`, and bumps status
        to `partial` or `paid` as appropriate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, minimum: 0.01 }
                payment_method: { type: string }
                reference: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Recorded }
        '400': { description: Invalid amount }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/transactions/invoice/{invoiceId}:
    parameters:
      - { in: path, name: invoiceId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [transactions]
      summary: List transactions linked to one invoice
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Transaction' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Must be a valid v4-style UUID; malformed values return 400.
  responses:
    BadRequest:
      description: Malformed request body
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
    BadUuid:
      description: Path param was not a UUID
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string, example: 'Invalid id: must be a UUID' } } }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
    NotFound:
      description: Entity not found (or belongs to a different tenant)
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        first_name: { type: string }
        last_name: { type: string }
        company_name: { type: string }
        vat_number: { type: string }
        phone: { type: string }
        address: { type: object, additionalProperties: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: JWT, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
    UserProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
        company_name: { type: string, nullable: true }
        vat_number: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: object, nullable: true, additionalProperties: true }
    Client:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, nullable: true }
        company_name: { type: string, nullable: true }
        vat_number: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: object, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ClientInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string }
        company_name: { type: string }
        vat_number: { type: string }
        phone: { type: string }
        address: { type: object }
    ClientList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Client' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    LineItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        description: { type: string }
        quantity: { type: string, example: '1.0000' }
        unit_price: { type: string, example: '1500.00' }
        amount: { type: string, example: '1500.00' }
        sort_order: { type: integer }
    LineItemInput:
      type: object
      required: [description, unit_price]
      properties:
        description: { type: string }
        quantity: { type: number, default: 1 }
        unit_price: { type: number }
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        client_id: { type: string, format: uuid, nullable: true }
        invoice_number: { type: string, example: INV-0001 }
        status: { type: string, enum: [draft, sent, partial, paid] }
        issue_date: { type: string, format: date, nullable: true }
        due_date: { type: string, format: date, nullable: true }
        sent_at: { type: string, format: date-time, nullable: true }
        paid_at: { type: string, format: date-time, nullable: true }
        subtotal: { type: string }
        tax_rate: { type: string }
        tax_amount: { type: string }
        discount_rate: { type: string }
        discount_amount: { type: string }
        retention_rate: { type: string }
        retention_amount: { type: string }
        cis_rate: { type: string }
        cis_amount: { type: string }
        total_amount: { type: string }
        amount_paid: { type: string }
        amount_due: { type: string }
        notes: { type: string, nullable: true }
        terms: { type: string, nullable: true }
    InvoiceFull:
      allOf:
        - $ref: '#/components/schemas/Invoice'
        - type: object
          properties:
            line_items: { type: array, items: { $ref: '#/components/schemas/LineItem' } }
            transactions: { type: array, items: { $ref: '#/components/schemas/Transaction' } }
    InvoiceInput:
      type: object
      properties:
        client_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, sent, partial, paid] }
        issue_date: { type: string, format: date }
        due_date: { type: string, format: date }
        tax_rate: { type: number, description: percent, example: 20 }
        discount_rate: { type: number, description: percent, example: 0 }
        retention_rate: { type: number, description: percent }
        cis_rate: { type: number, description: percent }
        invoice_prefix: { type: string, default: INV }
        auto_increment: { type: boolean, default: true }
        notes: { type: string }
        terms: { type: string }
        line_items: { type: array, items: { $ref: '#/components/schemas/LineItemInput' } }
    InvoiceList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Invoice' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        invoice_id: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [expense, payment] }
        amount: { type: string }
        transaction_date: { type: string, format: date }
        category: { type: string, nullable: true }
        description: { type: string, nullable: true }
        payment_method: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        notes: { type: string, nullable: true }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }